Rename VideoEmbed to VideoPreview and flatten Embed directory
The component renders an uploaded video attachment (poster, play badge,
duration) and opens the media viewer; it is not an external-content
embed, and it was the last file left in the Embed directory.
