--- packages/app/ui/components/Embed/index.tsx
@@ -1,3 +1,0 @@
-import VideoEmbed from './VideoEmbed';
-
-export { VideoEmbed };
